refactor: Extract interface for LocationRepository
